requirements.txt
numpy
matplotlib
torch
open3d
pillow
torchvision
kaleido